// backend/src/routes/auth/dto.ts
import { z } from "zod";

export const loginSchema = z.object({
    username: z.string().min(1, "Username is required"),
    password: z.string().min(1, "Password is required"),
});

export type LoginRequest = z.infer<typeof loginSchema>;

export type LoginResponse = {
    token: string;
    username: string;
};